last third commit --- add Seach emails and Weather tests and TestCases

+++ Svyrydiuk_G8_Exam_Project/src/main/java/pages/CurrencyExchangeRatePage.java
@@ -50,7 +50,7 @@ public class CurrencyExchangeRatePage extends ParentPage {
         enterTextIntoInput(amountCurrencyToSell, amount);
     }
 
-    public void checkExchangeCalculation(float amount) {
+    public void checkExchangeCalculation(double amount) {
         enterAmountCurrencyToSell(String.valueOf(amount));
         double sellUSD = getBuyUSD() * amount;
         logger.info("Sell USD = " + sellUSD);
